Added portlet table
