+++ src/Scenes/Generic-Listing/genericListing.scss
@@ -8,7 +8,8 @@
     .page-bar {
         height: 40px;
         display: flex;
-        justify-content: space-around;
+        justify-content: space-between;
+        width: 95%;
         padding: 4px 6px;
         background: #f1f1f1;
         border-bottom: 1px solid #d4d4d4;
