+++ src/Scenes/Header/header.scss
@@ -8,6 +8,7 @@
     .header-content {
         display: flex;
         justify-content: space-between;
-        flex-basis: 100%;
-    }
+        // flex-basis: 100%;
+
+        flex-basis: 96%;    }
 }
